requirements.txt
scrapy
python-box
